DynDEM: Split out memory circuit core functions
