+++ libs/qec/lib/extended_dem.cpp
@@ -326,14 +326,8 @@ extended_dem extended_dem_from_css_matrices(const css_code_matrices &code,
 } // end - extended_dem_from_css_matrices()
 
 // ---------------------------------------------------------------------------
-// dem_chunk_spec / dem_chunks_spec
+// dem_chunk_spec
 // ---------------------------------------------------------------------------
-
-// Number of rows a -1-terminated index list describes. validate() has already
-// established that the list ends with a terminator, so this is exact.
-static uint32_t sparse_row_count(const std::vector<std::int64_t> &rows) {
-  return static_cast<uint32_t>(std::count(rows.begin(), rows.end(), -1));
-}
 
 // Split a -1-terminated index list into a sparse matrix with num_faults
 // columns. Empty input yields a 0-row matrix that still carries the column
@@ -409,82 +403,8 @@ void dem_chunk_spec::validate(const std::string &context) const {
   validate_index_list(O_sparse, num_faults, context, "O_sparse");
 }
 
-bool dem_chunks_spec::is_empty() const {
-  return init.is_empty() && bulk.is_empty() && final.is_empty();
-}
-
-bool dem_chunks_spec::has_bulk() const { return !bulk.is_empty(); }
-
-void dem_chunks_spec::validate() const {
-  if (is_empty())
-    throw std::invalid_argument("dem_chunks: no phases were supplied");
-  if (init.is_empty())
-    throw std::invalid_argument("dem_chunks: init phase is required");
-  if (final.is_empty())
-    throw std::invalid_argument("dem_chunks: final phase is required");
-
-  init.validate("dem_chunks.init");
-  if (has_bulk())
-    bulk.validate("dem_chunks.bulk");
-  final.validate("dem_chunks.final");
-
-  // init starts the stream, so nothing feeds its incoming seam; final ends it,
-  // so nothing consumes its outgoing seam. init's own round-0 detectors belong
-  // in H_mid_sparse, compared against the zero initial state.
-  if (!init.H_in_sparse.empty())
-    throw std::invalid_argument(
-        "dem_chunks.init.H_in_sparse must be empty: the init phase has no "
-        "incoming seam. Put the first round's detectors in H_mid_sparse.");
-  if (!final.H_out_sparse.empty())
-    throw std::invalid_argument(
-        "dem_chunks.final.H_out_sparse must be empty: the final phase has no "
-        "outgoing seam.");
-
-  const uint32_t init_out = sparse_row_count(init.H_out_sparse);
-  const uint32_t final_in = sparse_row_count(final.H_in_sparse);
-
-  if (has_bulk()) {
-    const uint32_t bulk_in = sparse_row_count(bulk.H_in_sparse);
-    const uint32_t bulk_out = sparse_row_count(bulk.H_out_sparse);
-    // bulk is the phase that repeats, so it has to stitch to itself.
-    if (bulk_in != bulk_out)
-      throw std::invalid_argument(
-          "dem_chunks.bulk seams are unequal (H_in_sparse has " +
-          std::to_string(bulk_in) + " rows, H_out_sparse has " +
-          std::to_string(bulk_out) +
-          " rows); the repeated phase must stitch to itself");
-    if (init_out != bulk_in)
-      throw std::invalid_argument(
-          "dem_chunks seam width mismatch: init.H_out_sparse has " +
-          std::to_string(init_out) + " rows but bulk.H_in_sparse has " +
-          std::to_string(bulk_in));
-    if (bulk_out != final_in)
-      throw std::invalid_argument(
-          "dem_chunks seam width mismatch: bulk.H_out_sparse has " +
-          std::to_string(bulk_out) + " rows but final.H_in_sparse has " +
-          std::to_string(final_in));
-  } else if (init_out != final_in) {
-    throw std::invalid_argument(
-        "dem_chunks seam width mismatch: init.H_out_sparse has " +
-        std::to_string(init_out) + " rows but final.H_in_sparse has " +
-        std::to_string(final_in));
-  }
-
-  const uint32_t init_obs = sparse_row_count(init.O_sparse);
-  if (has_bulk() && sparse_row_count(bulk.O_sparse) != init_obs)
-    throw std::invalid_argument(
-        "dem_chunks observable count mismatch: init has " +
-        std::to_string(init_obs) + " but bulk has " +
-        std::to_string(sparse_row_count(bulk.O_sparse)));
-  if (sparse_row_count(final.O_sparse) != init_obs)
-    throw std::invalid_argument(
-        "dem_chunks observable count mismatch: init has " +
-        std::to_string(init_obs) + " but final has " +
-        std::to_string(sparse_row_count(final.O_sparse)));
-} // end - dem_chunks_spec::validate()
-
 // Build an extended_dem from a spec without validating. The caller must have
-// already called spec.validate() or dem_chunks_spec::validate().
+// already called spec.validate().
 static extended_dem dem_chunk_from_spec_impl(const dem_chunk_spec &spec) {
   extended_dem dem;
   dem.interior =
@@ -511,35 +431,6 @@ extended_dem dem_chunk_from_spec(const dem_chunk_spec &spec,
   spec.validate(context);
   return dem_chunk_from_spec_impl(spec);
 } // end - dem_chunk_from_spec()
-
-std::vector<extended_dem> dem_chunks_from_spec(const dem_chunks_spec &spec,
-                                               std::size_t num_rounds) {
-  if (num_rounds < 2)
-    throw std::invalid_argument(
-        "dem_chunks_from_spec: num_rounds must be at least 2 (init and final), "
-        "got " +
-        std::to_string(num_rounds));
-  spec.validate();
-
-  const std::size_t bulk_repeats = num_rounds - 2;
-  if (bulk_repeats > 0 && !spec.has_bulk())
-    throw std::invalid_argument("dem_chunks_from_spec: num_rounds " +
-                                std::to_string(num_rounds) + " needs " +
-                                std::to_string(bulk_repeats) +
-                                " bulk rounds but no bulk phase was supplied");
-
-  // spec.validate() already checked each phase; use the impl variant to avoid
-  // re-running per-phase validation for every call below.
-  std::vector<extended_dem> chunks;
-  chunks.reserve(num_rounds);
-  chunks.push_back(dem_chunk_from_spec_impl(spec.init));
-  if (bulk_repeats > 0) {
-    const auto bulk = dem_chunk_from_spec_impl(spec.bulk);
-    chunks.insert(chunks.end(), bulk_repeats, bulk);
-  }
-  chunks.push_back(dem_chunk_from_spec_impl(spec.final));
-  return chunks;
-} // end - dem_chunks_from_spec()
 
 // ---------------------------------------------------------------------------
 // dem_stitch
@@ -806,45 +697,6 @@ dem_chunks_to_detector_round(const std::vector<extended_dem> &dem_chunks) {
       result.insert(result.end(), d, round);
   }
   return result;
-}
-
-// dem_chunks_to_d_sparse: D_sparse[det_id] = measurement bit positions that
-// XOR-combine to form that detector.
-//   det k       (r=0): {k}               — compared to zero initial state
-//   det r*d+k (r>0): {(r-1)*d+k, r*d+k} — consecutive syndrome XOR
-std::vector<std::vector<uint32_t>>
-dem_chunks_to_d_sparse(const std::vector<extended_dem> &dem_chunks) {
-  validate_dem_chunk_sequence(dem_chunks, "dem_chunks_to_d_sparse");
-
-  const uint32_t d = chunk_seam_width(dem_chunks[0]);
-  // Rounds, not chunks: a chunk may carry several of them.
-  const std::size_t T = dem_chunks_to_rounds(dem_chunks);
-  const std::size_t n_det = T * static_cast<std::size_t>(d);
-
-  std::vector<std::vector<uint32_t>> d_sparse(n_det);
-
-  // Largest measurement bit index is (T-1)*d + (d-1); check it fits uint32_t.
-  if (T > 0) {
-    const auto max_bit = static_cast<std::size_t>(T - 1) * d + (d - 1);
-    if (max_bit > std::numeric_limits<uint32_t>::max())
-      throw std::overflow_error(
-          "dem_chunks_to_d_sparse: measurement bit index exceeds uint32_t max");
-  }
-
-  for (std::size_t r = 0; r < T; ++r) {
-    for (uint32_t k = 0; k < d; ++k) {
-      const std::size_t det = r * d + k;
-      if (r == 0) {
-        // Round-0 detector: single measurement at position k.
-        d_sparse[det] = {k};
-      } else {
-        // Later detector: XOR of previous and current round's measurement.
-        d_sparse[det] = {static_cast<uint32_t>((r - 1) * d + k),
-                         static_cast<uint32_t>(r * d + k)};
-      }
-    }
-  }
-  return d_sparse;
 }
 
 // dem_chunks_to_o_sparse: O_sparse[obs_id] = global fault column indices that
