+++ libs/qec/unittests/test_extended_dem.cpp
@@ -20,6 +20,7 @@
 //   - Per-qubit noise propagates correctly through stitch
 
 #include "cudaq/qec/code_matrices.h"
+#include "cudaq/qec/dem_chunks_memory.h"
 #include "cudaq/qec/dem_construction.h"
 #include "cudaq/qec/detector_error_model.h"
 #include "cudaq/qec/extended_dem.h"
